requirements.txt
streamlit>=1.20.0
langchain>=0.1.0
web3>=6.0.0
py-solc-x>=1.2.0
requests
python-dotenv
sseclient-py
langchain-core-1.0.5
